% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnorad.R
\name{ggnorad}
\alias{ggnorad}
\title{Wrapper function to load norad theme}
\usage{
ggnorad(style = "default")
}
\arguments{
\item{style}{Plot style. One of "default", "biglight1", "biglight2", "bigdark".}
}
\description{
Wrapper function to load norad theme
}
