% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norad_plot_palette.R
\docType{data}
\name{norad_colors}
\alias{norad_colors}
\title{Norad colors}
\format{
An object of class \code{character} of length 10.
}
\usage{
norad_colors
}
\description{
Norad colors
}
\keyword{datasets}
