% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{norad_gradient}
\alias{norad_gradient}
\title{Gradient color scale constructor for Norad colors}
\usage{
norad_gradient(n)
}
\arguments{
\item{n}{Number of colors}
}
\description{
Gradient color scale constructor for Norad colors
}
