% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norad_plot_palette.R
\name{norad_pal}
\alias{norad_pal}
\title{Return function to interpolate a Norad color palette}
\usage{
norad_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in norad_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a Norad color palette
}
