% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norad_plot_palette.R
\docType{data}
\name{norad_palettes}
\alias{norad_palettes}
\title{Combine colors into palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
norad_palettes
}
\description{
Combine colors into palettes
}
\keyword{datasets}
