% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{palette_gen}
\alias{palette_gen}
\title{Helper function for discrete color scales}
\usage{
palette_gen(palette = "main", direction = 1)
}
\arguments{
\item{palette}{Character name of palette in norad_palettes}

\item{direction}{Integer indicating whether the palette should be reversed}
}
\description{
Helper function for discrete color scales
}
