% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{palette_gen_c}
\alias{palette_gen_c}
\title{Helper function for continuous color scales}
\usage{
palette_gen_c(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{Character name of palette in norad_palettes}

\item{direction}{Integer indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to ggplot2::colorRampPalette()}
}
\description{
Helper function for continuous color scales
}
