% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{scale_color_norad_c}
\alias{scale_color_norad_c}
\title{Continious color scale constructor for Norad colors}
\usage{
scale_color_norad_c(palette = "greenyellow", direction = 1, ...)
}
\arguments{
\item{palette}{Character name of palette in norad_palettes}

\item{direction}{Integer indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
Continious color scale constructor for Norad colors
}
