% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{scale_fill_norad}
\alias{scale_fill_norad}
\title{Discrete fill scale constructor for Norad colors}
\usage{
scale_fill_norad(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{Character name of palette in norad_palettes}

\item{direction}{Integer indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to ggplot2::discrete_scale()}
}
\description{
Discrete fill scale constructor for Norad colors
}
