% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{scale_fill_norad_c}
\alias{scale_fill_norad_c}
\title{Continuous fill scale constructor for Norad colors}
\usage{
scale_fill_norad_c(palette = "greens", direction = 1, ...)
}
\arguments{
\item{palette}{Character name of palette in norad_palettes}

\item{direction}{Integer indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to ggplot2::scale_fill_gradientn()}
}
\description{
Continuous fill scale constructor for Norad colors
}
