% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{theme_norad_big_dark}
\alias{theme_norad_big_dark}
\title{Function for using Norad's theme in ggplot2 graphics}
\usage{
theme_norad_big_dark()
}
\description{
Function for using Norad's theme in ggplot2 graphics
}
