% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{theme_norad_map_light}
\alias{theme_norad_map_light}
\title{Function for using Norad's theme in ggplot2 graphics}
\usage{
theme_norad_map_light()
}
\description{
Function for using Norad's theme in ggplot2 graphics
}
