% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnorad.R
\name{ggnorad}
\alias{ggnorad}
\title{Wrapper function to load norad theme}
\usage{
ggnorad(style = "default")
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Wrapper function to load norad theme
}
