% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{norad_cols}
\alias{norad_cols}
\title{Function to extract Norad colors as hex codes}
\usage{
norad_cols(...)
}
\arguments{
\item{...}{Character names of norad_colors}
}
\description{
Function to extract Norad colors as hex codes
}
