% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{norad_gradient}
\alias{norad_gradient}
\title{Gradient color scale constructor for Norad colors}
\usage{
norad_gradient(n, color = norad_cols("green"))
}
\arguments{
\item{n}{Number of colors}

\item{color}{color upon which to generate gradient.}
}
\description{
Gradient color scale constructor for Norad colors
}
