% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{norad_pal}
\alias{norad_pal}
\title{Function to access Norad palettes}
\usage{
norad_pal(palette = "main", ...)
}
\arguments{
\item{palette}{Character name of palette in norad_palettes}

\item{...}{Additional arguments (not in use)}
}
\description{
Function to access Norad palettes
}
